Add a Kind type that is the union of all concrete types
This allows you to do exhaustiveness checking on type-level programming,
where you continually narrow the type via successive `if` statements
checking `instanceof` concrete types, and with a final `else` that
assumes your last type is the last one you've left to enumerate. An
example of this can be found in the `test/kind.ts` file.