Somewhat hacky generic support for GetType
Unfortunately due to the way the TS type system works --no rank-n
polymorphism -- we can't have functions that maintain their generic-ness
when passed around. So with GetType on functions, the generic types all
get set to `{}` which isn't ideal. But it's better than just making the
whole thing a `never`.