Make optional() keys TypeScript-compatible
The previous implementation of optional() broke compatibility with
TypeScript's type system; in TS, optional keys may either be left out,
or set to undefined. There isn't a built-in way to specify a key can be
left out *but not set to undefined* in a structural type, outside of
fairly awkward conditional type shenanigans; the built-in default ?
syntax allows undefined.

However, Structural's built-in default of `optional(...)` did the
opposite: it allowed the key to be left out, but failed if the key was
present and the value was undefined. This was surprising, in that it
differs from TS and Structural claims compatibility, and it was also
painful for TS generation, since it meant that we'd generate
slightly-invalid TypeScript if we tried to use optional keys for the
Partials. To work around this, Structural's `partial(...)` family of
functions *didn't* simply use `optional(...)`; it also applied a union
to undefined, so that Structural's partial acted like TypeScript's
Partial. But then the codegen was super awkward, since by default it
kept the union type in, which is unnecessary bloat for TS since TS
allows the union by default with just the ? syntax.

Rather than continue patching around our incompatibility, this commit
makes `optional(...)` act like ? and accept undefined. If you really
want the previous behavior, you can use the new `allowMissing(...)`,
which uses the previous default behavior of allowing missing keys but
failing on undefined.